backend/authentication-service/requirements.txt
# Self-contained for Docker.
# Local dev: pip install -r ../requirements.txt -r requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
pyjwt>=2.8.0
redis>=5.0.3
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1            # passlib is incompatible with bcrypt>=4.1 (72-byte limit)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# Testing
pytest>=8.1.0
httpx>=0.27.0
pytest-asyncio>=0.23.0
